Update game rating to IntegerField and enhance UI
Changed the Game model's rating field from FloatField to IntegerField for consistency and updated the corresponding migration. Improved the game list UI with a card flip effect, richer details, and visual enhancements. Added custom Tailwind keyframes for shimmer animation in base.html.

+++ whichgame/models.py
@@ -13,7 +13,7 @@ class Game(models.Model):
     genres = models.JSONField(default=list)    # ex: ["RPG", "Action"]
     
     # Les stats
-    rating = models.FloatField(null=True, blank=True)      
+    rating = models.IntegerField(null=True, blank=True)      
     price_current = models.DecimalField(max_digits=6, decimal_places=2, null=True, blank=True) 
     playtime_main = models.IntegerField(null=True, blank=True) # En heures
